changed server folder structure and added myself to codeowner
