+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-*   @jnwong116 @marianmwang @calebji123
+*   @jnwong116 @marianmwang @calebji123 @neolou
